Rearrange the files in SimG4CMS/ShowerLibraryProducer and collapse .h and .cc files
