+++ SimG4CMS/ShowerLibraryProducer/plugins/HcalForwardAnalysis.cc
@@ -2,26 +2,95 @@
 #include <cmath>
 #include <iostream>
 #include <iomanip>
+#include <string>
+#include <vector>
 
 // user include files
-#include "SimG4CMS/ShowerLibraryProducer/interface/HcalForwardAnalysis.h"
-
-#include "SimG4Core/Notification/interface/BeginOfRun.h"
-#include "SimG4Core/Notification/interface/BeginOfEvent.h"
-#include "SimG4Core/Notification/interface/EndOfEvent.h"
-
 #include "DataFormats/Math/interface/Point3D.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
+y
+#include "SimG4Core/Notification/interface/BeginOfRun.h"
+#include "SimG4Core/Notification/interface/BeginOfEvent.h"
+#include "SimG4Core/Notification/interface/EndOfEvent.h"
+#include "SimG4Core/Watcher/interface/SimProducer.h"
+#include "SimG4Core/Notification/interface/Observer.h"
 
-#include "G4SDManager.hh"
-#include "G4VProcess.hh"
+#include "SimG4CMS/ShowerLibraryProducer/interface/FiberG4Hit.h"
+#include "SimG4CMS/ShowerLibraryProducer/interface/HFShowerG4Hit.h"
+
 #include "G4HCofThisEvent.hh"
+#include "G4SDManager.hh"
+#include "G4Step.hh"
+#include "G4Track.hh"
+#include "G4ThreeVector.hh"
+#include "G4VProcess.hh"
 #include "CLHEP/Units/GlobalSystemOfUnits.h"
 #include "CLHEP/Units/GlobalPhysicalConstants.h"
+
+#include "TFile.h"
+#include "TTree.h"
+
+#define EDM_ML_DEBUG
+
+class HcalForwardAnalysis : public SimProducer,
+                            public Observer<const BeginOfRun*>,
+                            public Observer<const BeginOfEvent*>,
+                            public Observer<const EndOfEvent*>,
+                            public Observer<const G4Step*> {
+public:
+  struct Photon {
+    Photon(int id, float X, float Y, float Z, float T, float Lambda)
+        : fiberId(id), x(X), y(Y), z(Z), t(T), lambda(Lambda) {}
+    int fiberId;
+    float x;
+    float y;
+    float z;
+    float t;
+    float lambda;
+  };
+
+  HcalForwardAnalysis(const edm::ParameterSet& p);
+  HcalForwardAnalysis(const HcalForwardAnalysis&) = delete;  // stop default
+  const HcalForwardAnalysis& operator=(const HcalForwardAnalysis&) = delete;
+  ~HcalForwardAnalysis() override;
+
+  void produce(edm::Event&, const edm::EventSetup&) override;
+
+private:
+  void init();
+
+  // observer methods
+  void update(const BeginOfRun* run) override;
+  void update(const BeginOfEvent* evt) override;
+  void update(const G4Step* step) override;
+  void update(const EndOfEvent* evt) override;
+  //  void write(const EndOfRun * run);
+
+  //User methods
+  void setPhotons(const EndOfEvent* evt);
+  //void fillEvent(PHcalForwardLibInfo&);
+  void fillEvent();
+  void parseDetId(int id, int& tower, int& cell, int& fiber);
+  void clear();
+
+  edm::Service<TFileService> theFile;
+  TTree* theTree;
+  int theEventCounter;
+  int count;
+  int evNum;
+  float x[10000], y[10000], z[10000], t[10000], lambda[10000];
+  float primX, primY, primZ, primT;
+  float primMomX, primMomY, primMomZ;
+  int nphot;
+  int fiberId[10000];
+  std::vector<Photon> thePhotons;
+  std::vector<std::string> theNames;
+  bool fillt;
+};
 
 HcalForwardAnalysis::HcalForwardAnalysis(const edm::ParameterSet& p) {
   edm::ParameterSet m_SLP = p.getParameter<edm::ParameterSet>("HFShowerLibraryProducer");
@@ -85,7 +154,9 @@ void HcalForwardAnalysis::update(const BeginOfRun* run) {
 void HcalForwardAnalysis::update(const BeginOfEvent* evt) {
   evNum = (*evt)()->GetEventID();
   clear();
+#ifdef EDM_ML_DEBUG
   edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis: =====> Begin of event = " << evNum;
+#endif
 }
 
 void HcalForwardAnalysis::update(const G4Step* aStep) {}
@@ -94,7 +165,9 @@ void HcalForwardAnalysis::update(const EndOfEvent* evt) {
   count++;
 
   //fill the buffer
+#ifdef EDM_ML_DEBUG
   edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis::Fill event " << (*evt)()->GetEventID();
+#endif
   setPhotons(evt);
 
   int iEvt = (*evt)()->GetEventID();
@@ -114,6 +187,7 @@ void HcalForwardAnalysis::setPhotons(const EndOfEvent* evt) {
   FiberG4HitsCollection* theHC;
   // Look for the Hit Collection of HCal
   G4HCofThisEvent* allHC = (*evt)()->GetHCofThisEvent();
+#ifdef EDM_ML_DEBUG
   edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis:: Has " << allHC->GetNumberOfCollections()
                                      << " collections";
   for (int k = 0; k < allHC->GetNumberOfCollections(); ++k) {
@@ -122,22 +196,29 @@ void HcalForwardAnalysis::setPhotons(const EndOfEvent* evt) {
     edm::LogVerbatim("HcalForwardLib") << "Collecttion[" << k << "] " << allHC->GetHC(k) << "  " << name << ":"
                                        << nameSD;
   }
+#endif
   std::string sdName = theNames[0];  //name for fiber hits
   idHC = G4SDManager::GetSDMpointer()->GetCollectionID(sdName);
   theHC = (FiberG4HitsCollection*)allHC->GetHC(idHC);
+#ifdef EDM_ML_DEBUG
   edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis::setPhotons() Hit Collection for " << sdName << " of ID "
                                      << idHC << " is obtained at " << theHC;
+#endif
   std::vector<HFShowerPhoton> ShortFiberPhotons;
   std::vector<HFShowerPhoton> LongFiberPhotons;
   LongFiberPhotons.clear();
   ShortFiberPhotons.clear();
   if (idHC >= 0 && theHC != nullptr) {
     int thehc_entries = theHC->entries();
+#ifdef EDM_ML_DEBUG
     edm::LogVerbatim("HcalForwardLib") << "FiberhitSize " << thehc_entries;
+#endif
     for (j = 0; j < thehc_entries; j++) {
       FiberG4Hit* aHit = (*theHC)[j];
       std::vector<HFShowerPhoton> thePhotonsFromHit = aHit->photon();
+#ifdef EDM_ML_DEBUG
       edm::LogVerbatim("HcalForwardLib") << "Fiberhit " << j << " has " << thePhotonsFromHit.size() << " photons.";
+#endif
       int fTowerId = -1;
       int fCellId = -1;
       int fFiberId = -1;
@@ -148,13 +229,17 @@ void HcalForwardAnalysis::setPhotons(const EndOfEvent* evt) {
         if (aHit->depth() == 2)
           ShortFiberPhotons.push_back(thePhotonsFromHit[iph]);
       }
+#ifdef EDM_ML_DEBUG
       edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis::setPhotons() NbPhotons " << thePhotonsFromHit.size()
                                          << " towerId " << fTowerId << " cellId " << fCellId << " fiberId " << fFiberId
                                          << " depth " << aHit->depth();
+#endif
     }
   } else {
     fillt = false;
+#ifdef EDM_ML_DEBUG
     edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis::setPhotons(): No Photons!";
+#endif
     return;
   }
   edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis::setPhotons() LongFibPhotons: " << LongFiberPhotons.size()
@@ -181,9 +266,11 @@ void HcalForwardAnalysis::setPhotons(const EndOfEvent* evt) {
     int thec_hc_entries = theChamberHC->entries();
     for (j = 0; j < thec_hc_entries; ++j) {
       HFShowerG4Hit* aHit = (*theChamberHC)[j];
+#ifdef EDM_ML_DEBUG
       edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis::setPhotons() Chamber Hit id " << aHit->hitId()
                                          << " track id " << aHit->trackId() << " prim. pos. " << aHit->globalPosition()
                                          << " prom mom. dir. " << aHit->primaryMomDir() << " time " << aHit->time();
+#endif
       primPosOnSurf.SetXYZ(aHit->globalPosition().x(), aHit->globalPosition().y(), aHit->globalPosition().z());
       primMomDirOnSurf.SetXYZ(aHit->primaryMomDir().x(), aHit->primaryMomDir().y(), aHit->primaryMomDir().z());
       primTimeOnSurf = aHit->time();
@@ -251,9 +338,9 @@ void HcalForwardAnalysis::setPhotons(const EndOfEvent* evt) {
 }
 
 void HcalForwardAnalysis::fillEvent() {
-  /*
+#ifdef EDM_ML_DEBUG
     edm::LogVerbatim("HcalForwardLib") << "HcalForwardAnalysis: =====> filledEvent";
-  */
+#endif
   nphot = int(thePhotons.size());
   for (int i = 0; i < nphot; ++i) {
     x[i] = thePhotons[i].x;
@@ -287,3 +374,8 @@ void HcalForwardAnalysis::clear() {
 
   thePhotons.clear();
 }
+
+#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+
+DEFINE_SIMWATCHER(HcalForwardAnalysis);
