--- SimG4CMS/ShowerLibraryProducer/plugins/module.cc
@@ -1,7 +1,0 @@
-#include "SimG4CMS/ShowerLibraryProducer/interface/HcalForwardAnalysis.h"
-#include "SimG4CMS/ShowerLibraryProducer/interface/CastorShowerLibraryMaker.h"
-#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-
-DEFINE_SIMWATCHER(HcalForwardAnalysis);
-DEFINE_SIMWATCHER(CastorShowerLibraryMaker);
